httpdate: fix setting the date, 3rd try
Both implementations of date command from GNU and from BusyBox
take an argument for setting the time of @<timestamp> format.
This format is timezone independent, because it is measuring the
time from `1970-01-01 00:00:00 UTC` so it's always measured in
UTC. Set remote and local datetime objects to UTC timezone, to
make sure there will be no surprises with delta offset. Hardcode
GMT timezone, as that is the only format httpdate supports in
HTTP headers.